[Spring, Boot, Starter, Actuator] implemented HW43

+++ src/main/java/com/tms/lesson43/hw/controller/UserController.java
@@ -1,5 +1,7 @@
 package com.tms.lesson43.hw.controller;
 
+import java.util.Objects;
+
 import com.tms.lesson43.hw.model.User;
 import com.tms.lesson43.hw.service.UserService;
 import org.springframework.stereotype.Controller;
@@ -14,6 +16,8 @@ import org.springframework.ui.Model;
 @RequestMapping("/user")
 public class UserController {
 
+    String errorMessage = "Please, fill in valid data.";
+
     @GetMapping(value = "/create")
     public ModelAndView fillCreateUserForm() {
         return new ModelAndView("create/createUserForm", "user", new User());
@@ -21,15 +25,40 @@ public class UserController {
 
     @PostMapping(value = "/create")
     public ModelAndView createUser(@ModelAttribute User user, Model model) {
-        if (user.getId() != 0 && !user.getPassport_number().isEmpty()) {
+        if (Objects.isNull(UserService.getUserInfo(user.getId())) &&
+                user.getId() != 0 &&
+                !user.getName().isEmpty() &&
+                !user.getSurname().isEmpty() &&
+                user.getAge() != 0 &&
+                !user.getPassport_number().isEmpty()) {
             UserService.createUser(user);
             String message = String.format("User with '%s' id has been created successfully!", user.getId());
             model.addAttribute("message", message);
             return new ModelAndView("create/createUser", "user", new User());
         }
-        String message = "Please, fill in all required data.";
-        model.addAttribute("message", message);
+        model.addAttribute("message", errorMessage);
         return new ModelAndView("create/createUserForm", "user", new User());
+    }
+
+    @GetMapping(value = "/get")
+    public ModelAndView fillGetUserForm() {
+        return new ModelAndView("get/getUserForm", "user", new User());
+    }
+
+    @PostMapping(value = "/get")
+    public ModelAndView getUser(@ModelAttribute User user, Model model) {
+        if (!Objects.isNull(UserService.getUserInfo(user.getId())) && user.getId() != 0) {
+            user = UserService.getUserInfo(user.getId());
+            model.addAttribute("id", user.getId());
+            model.addAttribute("name", user.getName());
+            model.addAttribute("surname", user.getSurname());
+            model.addAttribute("age", user.getAge());
+            model.addAttribute("passport_number", user.getPassport_number());
+            return new ModelAndView("get/getUser", "user", new User());
+
+        }
+        model.addAttribute("message", errorMessage);
+        return new ModelAndView("get/getUserForm", "user", new User());
     }
 
     @GetMapping(value = "/update")
@@ -39,14 +68,13 @@ public class UserController {
 
     @PostMapping(value = "/update")
     public ModelAndView updateUser(@ModelAttribute User user, Model model) {
-        if (user.getId() != 0) {
+        if (!Objects.isNull(UserService.getUserInfo(user.getId())) && user.getId() != 0) {
             UserService.updateUser(user);
             String message = String.format("User with '%s' id has been updated successfully!", user.getId());
             model.addAttribute("message", message);
             return new ModelAndView("update/updateUser", "user", new User());
         }
-        String message = "Please, fill in all required data.";
-        model.addAttribute("message", message);
+        model.addAttribute("message", errorMessage);
         return new ModelAndView("update/updateUserForm", "user", new User());
     }
 
@@ -57,40 +85,13 @@ public class UserController {
 
     @PostMapping(value = "/delete")
     public ModelAndView deleteUser(@ModelAttribute User user, Model model) {
-        if (user.getId() != 0) {
+        if (!Objects.isNull(UserService.getUserInfo(user.getId())) && user.getId() != 0) {
             UserService.deleteUser(user);
             String message = String.format("User with '%s' id has been deleted successfully!", user.getId());
             model.addAttribute("message", message);
             return new ModelAndView("delete/deleteUser", "user", new User());
         }
-        String message = "Please, fill in all required data.";
-        model.addAttribute("message", message);
+        model.addAttribute("message", errorMessage);
         return new ModelAndView("delete/deleteUserForm", "user", new User());
-    }
-
-    @GetMapping(value = "/get")
-    public ModelAndView fillGetUserForm() {
-        return new ModelAndView("get/getUserForm", "user", new User());
-    }
-
-    @PostMapping(value = "/get")
-    public ModelAndView getUser(@ModelAttribute User user, Model model) {
-        if (user.getId() != 0) {
-            user = UserService.getUserInfo(user);
-            System.out.println("!!!!!!!!!!!hello");
-            System.out.println(user);
-            if (user != null) {
-                model.addAttribute("isUserExist", true);
-                model.addAttribute("id", user.getId());
-                model.addAttribute("name", user.getName());
-                model.addAttribute("surname", user.getSurname());
-                model.addAttribute("age", user.getAge());
-                model.addAttribute("passport_number", user.getPassport_number());
-                return new ModelAndView("get/getUser", "user", new User());
-            }
-        }
-        model.addAttribute("isUserExist", false);
-        model.addAttribute("message", "User with '" + user.getId() + "' ID doesn't exist.");
-        return new ModelAndView("get/getUserForm", "user", new User());
     }
 }
